+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - make SPI `write` method public
 - `set_temp_monitor` method
 - `set_lna_gain` method
+- `set_ocp`, `set_ocp_trim` methods
 
 ## [0.1.0] - 2026-03-07